backend/requirements.txt
appdirs==1.4.4
asgiref==3.5.0
astroid==2.4.2
Babel==2.9.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cmake==3.22.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.13
django-cors-headers==3.11.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.3
importlib-metadata==1.7.0
isort==5.7.0
itypes==1.2.0
Jinja2==3.1.1
lazy-object-proxy==1.4.3
lxml==4.6.3
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
passlib==1.7.4
Pillow==8.1.0
polib==1.1.1
psutil==5.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.6.0
PyPDF2==1.26.0
# pypiwin32==223
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2022.1
# pywin32==300
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
typed-ast==1.4.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.4.2
Werkzeug==1.0.1
whitenoise==6.1.0
wrapt==1.12.1
zipp==3.8.0
